refactor: ClubDetailedResponse 결과값을 club이란 객체로 wrap 하였음
