+++ backend/src/main/java/moadong/club/payload/dto/ClubDetailedResult.java
@@ -1,32 +1,33 @@
-package moadong.club.payload.response;
+package moadong.club.payload.dto;
 
 import lombok.Builder;
 import moadong.club.entity.Club;
 import moadong.club.entity.ClubInformation;
+import moadong.club.payload.response.ClubDetailedResponse;
 
 import java.util.List;
 
 @Builder
-public record ClubDetailedPageResponse(
-        String id,
-        String name,
-        String logo,
-        List<String> tags,
-        String state,
-        List<String> feeds,
-        String description,
-        String presidentName,
-        String presidentPhoneNumber,
-        String recruitmentPeriod,
-        String classification,
-        String division
-) {
-    public static ClubDetailedPageResponse createClubDetailedPageResponse(
+public record ClubDetailedResult(
+    String id,
+    String name,
+    String logo,
+    List<String> tags,
+    String state,
+    List<String> feeds,
+    String description,
+    String presidentName,
+    String presidentPhoneNumber,
+    String recruitmentPeriod,
+    String classification,
+    String division
+){
+    public static ClubDetailedResult of(
             Club club,
             ClubInformation clubInformation,
             List<String> clubFeedImages,
             List<String> clubTags) {
-        return ClubDetailedPageResponse.builder()
+        return ClubDetailedResult.builder()
                 .id(club.getId())
                 .name(club.getName())
                 .classification(club.getClassification())
@@ -39,4 +40,5 @@ public record ClubDetailedPageResponse(
                 .tags(clubTags)
                 .build();
     }
+
 }
