+++ backend/src/main/java/moadong/club/payload/response/ClubDetailedResponse.java
@@ -1,0 +1,10 @@
+package moadong.club.payload.response;
+
+import lombok.Builder;
+import moadong.club.payload.dto.ClubDetailedResult;
+
+@Builder
+public record ClubDetailedResponse(
+    ClubDetailedResult club
+) {
+}
